Fixed generation of invalid default arguments.
This fixes a bug where default arguments that were unresolved
could occur after default arguments that were resolved.  This
would lead to output like:

`_arg("a") = "foo", _arg("b"), _arg("c") = "bar"

which would not be a valid set of default arguments.  The fix
searches for the last unresolved argument in the parameter list,
and only provides defaults for parameters after this one.