Added flag to control docs extraction.
This adds a new flag `-no-docs` to control the automatic addition
of the `-Wdocumentation/-Wno-documentation` flags to the internal
compiler arguments used in Chimera.

This avoids having to do comment parsing, or pass the non-standard
`-Wdocumentation` flag, when doing test compilations to figure out
the necessary flags with which to run chimera.