Added fix for HTML escaping.
This commit adds a custom escaping function to the mstch namespace
that disables all HTML escaping.  Since HTML escaping heavily
interferes with generation of valid C++ type declarations, this
was causing numerous issues in code generation.

As a side effect, the _unescaped_ template include operator for
mstch, `{{{ }}}`, should no longer be necessary, so I have
removed it from all instances of templates in the source tree.

Closes #104.