Staged `sources.txt` generation in CMake binding helpers.
Previously, `sources.txt` was directly piped from the output of
chimera, but was not cleared or invalidated if the command did not
complete successfully, leading to subsequent build steps skipping
the regeneration of sources.

This PR changes the `add_custom_command()` directive to stage the
output of chimera to a tempfile first, which is moved into location
after the success of the generation command.  This prevents
incomplete or invalid `sources.txt` from being used by subsequent
build steps.

As a side effect of this PR, I discovered a bug in the unit test
CMakeLists.txt that was leading to the test practically always
succeeding because it did not have the `chimeraExecutable` variable
set and somehow managed to run a very ill defined bash command
that always succeeded.