Overhaul of YAML usage in chimera::CompiledConfiguration. (#135)
* Staged `sources.txt` generation in CMake binding helpers.

Previously, `sources.txt` was directly piped from the output of chimera, but was not cleared or invalidated if the command did not complete successfully, leading to subsequent build steps skipping the regeneration of sources.

This PR changes the `add_custom_command()` directive to stage the output of chimera to a tempfile first, which is moved into location after the success of the generation command.  This prevents incomplete or invalid `sources.txt` from being used by subsequent build steps.

As a side effect of this PR, I discovered a bug in the unit test CMakeLists.txt that was leading to the test practically always succeeding because it did not have the `chimeraExecutable` variable set and somehow managed to run a very ill defined bash command that always succeeded.

* Added YAML error checking around CompiledConfiguration.

This PR rewrites most of the YAML access in CompiledConfiguration to fix a number of parsing issues related to missing YAML entries, improve error checking and output, and generalize the text-template engine.

YAML accessors are all result-checked to eliminate many YAML::InvalidNode runtime errors. Default bindings are now represented via chimera::Binding structs, which are resolved once when the configuration is compiled. Templates can now use any hierarchy of parameters under the template::file and template:main namespaces. File references in these parameters are recursively resolved and automatically forwarded directly to mstch templates so we do not have to hardcode parameter names.

Our poor single unit test now actually generates cpp output and may actually be close to working now. I added Boost.Python as a CI dependency so that we can compile our unit test which generates a sort of python binding by default.

There are two minor breaking changes to the YAML configuration file:

1. To include file-snippets in the template section, instead of using a map with a key entry of source:, we check for a scalar with a type tag of !file. This disambiguates file loading from hierarchical parameters.

2. The template::cxx_record entry has been changed to template::class because it is a less unwieldy name.

Closes #120.