Modified CMake to use newer directives. (#142)
This PR updates the CMakeLists.txt to use newer `target_` directives for includes and compile options, which removes some of the cruft associated with inheriting these properties.

The required version of CMake has been updated to 2.8.12 to make use of these newer features. This should be available on all targeted platforms (and specifically, Ubuntu 14.04).