Support for multiple built-in language bindings. (#143)
This PR overhauls the binding class stubs added in #135 into a full template generation subsystem to package multiple language bindings. At the moment, these languages can be selected by defining a binding: [BINDING_NAME] key in the configuration YAML, but we can rearrange this as needed later.

Each binding is defined as a set of template files in a subdirectory of bindings/. The CMakeLists.txt in this directory iterates through a list of these template files, assembles them into -impl.h headers, and then includes those headers into a binding.cpp file which is used to build a static library of built-in bindings which is accessible via a global map in binding.h. These -impl.h files are then loaded in the constructor of the singleton Configuration, and queried in the constructor of CompiledConfiguration.