Fixes add_chimera_binding() to inherit target properties.
This commit fixes the issue originally raised in #113 where build
options for the chimera binding target were not propagated to the
placeholder target used in binding generation.

The fix makes use of generator expressions to directly insert the
properties for the target into the properties of the placeholder.
This works for compile options, include directories, and link
libraries, but I don't know if it works quite correctly with
transitive dependencies directly on other CMake targets.