Add error message when expected scalar is not found.
The `chimera::CompiledConfiguration::Lookup()` function is intended
to handle either scalar values or (when a `file` tag is specified)
a path to a file on disk.

This commit throws an error when neither of those two values is
found in the node provided to `Lookup()`, instead of the previous
behavior of simply returning an empty string instead.

This prevents configuration parsing issues from being masked by other
errors due to the empty string.