Added ordering to traversed namespaces in configuration.
This adds a bugfix where parent namespace declarations were being
added to the traversed namespace list of the `CompiledConfiguration`
in an arbitrary order, and thus provided to the binding template
in an order that could not be used to construct namespace hierarchical
objects.

Now, when a traversed namespace is added to a `CompiledConfiguration`,
the enclosing namespaces are first added, to ensure that they will
be provided to the binding template in a consistent order.