Found a much simpler way to do this.
Instead of doing a hierarchical traversal on each element, we
can note that the calling consumer guarantees a hierarchical
traversal order starting with the root namespaces.

We already depend on this traversal order in other places, so it
is reasonable to simply add the mechanism to preserve that order,
which was the actual bug in the first place.